Clean up ESP32 PWM handling
- Use of 13 bit resolution causes some unexpected behavior
- Simplify build dependencies
