+++ examples/ConfigurableFirmata/ConfigurableFirmata.ino
@@ -47,7 +47,6 @@ AnalogInputFirmata analogInput;
 
 #include <AnalogOutputFirmata.h>
 AnalogOutputFirmata analogOutput;
-#include <AnalogWrite.h>
 #endif
