+++ src/AnalogOutputFirmata.h
@@ -20,10 +20,6 @@
 #include <ConfigurableFirmata.h>
 #include "FirmataFeature.h"
 
-// analogWriteCallback is defined in AnalogWrite.h but must also be declared here in order
-// for AnalogOutputFirmata to compile
-void analogWriteCallback(byte pin, int value);
-
 class AnalogOutputFirmata: public FirmataFeature
 {
     friend void analogWrite(byte pin, uint32_t value);
@@ -44,4 +40,5 @@ class AnalogOutputFirmata: public FirmataFeature
 #endif
 };
 
+void analogWriteCallback(byte pin, int value);
 #endif
